Move Google Video code into another file
html.js is getting messy, and we'll be adding more Google Video stuff
